fix: support 64-bit person ids

+++ androidApp/src/androidMain/kotlin/com/continuum/app/android/ui/navigation/Routes.kt
@@ -98,7 +98,7 @@ sealed class Route(val route: String) {
         }
     }
 
-    data class PersonDetail(val personId: Int) : Route("person/$personId") {
+    data class PersonDetail(val personId: Long) : Route("person/$personId") {
         companion object {
             const val ROUTE = "person/{personId}"
         }
