+++ androidTvApp/src/androidMain/kotlin/com/continuum/app/tv/ui/screens/people/TvPersonDetailScreen.kt
@@ -72,7 +72,7 @@ import org.koin.core.parameter.parametersOf
  */
 @Composable
 fun TvPersonDetailScreen(
-    personId: Int,
+    personId: Long,
     onOpenItemDetail: (contentId: String) -> Unit,
     onBack: () -> Unit,
     viewModel: TvPersonDetailViewModel = koinViewModel(
